because deleted Conference+Extension, move testSchedulesFiltered to ScheduleTests
