+++ App/App/App.xctestplan
@@ -23,13 +23,6 @@
         "identifier" : "ScheduleFeatureTests",
         "name" : "ScheduleFeatureTests"
       }
-    },
-    {
-      "target" : {
-        "containerPath" : "container:..\/MyLibrary",
-        "identifier" : "SharedModelsTests",
-        "name" : "SharedModelsTests"
-      }
     }
   ],
   "version" : 1
