+++ MyLibrary/Package.swift
@@ -104,12 +104,5 @@ let package = Package(
         .product(name: "ComposableArchitecture", package: "swift-composable-architecture"),
       ]
     ),
-    .testTarget(
-      name: "SharedModelsTests",
-      dependencies: [
-        "SharedModels",
-        .product(name: "ComposableArchitecture", package: "swift-composable-architecture"),
-      ]
-    ),
   ]
 )
