+++ MyLibrary/Tests/ScheduleFeatureTests/ScheduleTests.swift
@@ -115,4 +115,16 @@ final class ScheduleTests: XCTestCase {
     initialState.favorites = [initialState.day1!.title: [firstSession]]
     return initialState
   }()
+
+  @MainActor
+  func testSchedulesFiltered() {
+    let schedulesWith2Sessions = [Schedule(time: Date(timeIntervalSince1970: 10_000), sessions: [.mock1, .mock2])]
+    let conference = Conference(id: 1, title: "conference", date: Date(timeIntervalSince1970: 1_000), schedules: schedulesWith2Sessions)
+    let favoritesMock1Only: Favorites = [conference.title: [.mock1]]
+
+    let actual = schedulesWith2Sessions.filtered(using: favoritesMock1Only, in: conference)
+
+    let schedulesMock1Only = [Schedule(time: Date(timeIntervalSince1970: 10_000), sessions: [.mock1])]
+    XCTAssertEqual(actual, schedulesMock1Only)
+  }
 }
