--- MyLibrary/Tests/SharedModelsTests/Mocks.swift
@@ -1,6 +1,0 @@
-import SharedModels
-
-extension Session {
-  static let mock1 = Session(title: "1", speakers: nil, place: nil, description: nil, requirements: nil)
-  static let mock2 = Session(title: "2", speakers: nil, place: nil, description: nil, requirements: nil)
-}
